add new metric module to include context, monitor and metric classes
